refactor server logic from top level app logic
